created .gitignore

+++ scr/helper_files/genome_to_molecule.py
@@ -425,10 +425,10 @@ def load_building_blocks(path):
 	dbPc5 = Building_Block(abbrev="dbPc5", num_atoms=58,origin=12, para_pos=26, para_angle=0, meta_pos=20 , meta_angle = -np.pi/3., ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=1, path=path+"/dbPc5.xyz")
 	pseudo_para_naph_PCP = Building_Block(abbrev="pseudo-para_naph_PCP", num_atoms=44,origin=0, para_pos=18, para_angle=0, meta_pos=16 , meta_angle = -np.pi/3, ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=1, path=path+"/pseudo-para_naph_PCP.xyz")
 	line =Building_Block(abbrev="line", num_atoms=4,origin=0, para_pos=1, para_angle=0, meta_pos=1 , meta_angle = 0., ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=1, path=path+"/line.xyz")
-	rot=Building_Block(abbrev="line", num_atoms=47,origin=6, para_pos=16, para_angle=0, meta_pos=20 , meta_angle = 0., ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=2, path=path+"/rot.xyz")
-	stacked_anth=Building_Block(abbrev="stacked_anth", num_atoms=62,origin=3, para_pos=22, para_angle=0, meta_pos=30 , meta_angle = 0., ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=2, path=path+"/stacked_anth.xyz")
+	#rot=Building_Block(abbrev="line", num_atoms=47,origin=6, para_pos=16, para_angle=0, meta_pos=20 , meta_angle = 0., ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=2, path=path+"/rot.xyz")
+	#stacked_anth=Building_Block(abbrev="stacked_anth", num_atoms=62,origin=3, para_pos=22, para_angle=0, meta_pos=30 , meta_angle = 0., ortho_pos=0, ortho_angle=-2.*np.pi/3, fixed_left = -1,complexity=2, path=path+"/stacked_anth.xyz")
 	
-	building_blocks = [benzene,napthtalene,dbPc1,dbPc4,dbPc6, dbPc5,pseudo_para_naph_PCP, line,rot,stacked_anth]
+	building_blocks = [benzene,napthtalene,dbPc1,dbPc4,dbPc6, dbPc5,pseudo_para_naph_PCP, line]
 	#building_blocks = [benzene,napthtalene]
 
 	return building_blocks
@@ -518,6 +518,9 @@ def process_genome(generation : int, individual: int, genome:Genome, run_path):
 			print("scr_dir " + str(scr_dir))
 			dst_dir = generation_data_path + "/" + str(generation)+ "/" +str(individual) + "/"
 			print("dst_dir " + str(dst_dir))
+			if(path.exists(dst_dir) == True):
+				print("Other job running ... Aborting")
+				raise ValueError('Other job running ... Aborting')
 			os.system("mkdir " + str(dst_dir))
 			dst_dir += "."
 			os.system("cp -R " + scr_dir + " " + dst_dir)
@@ -574,7 +577,7 @@ if __name__ == '__main__':
 
 	#construction_loop(genome, building_blocks, "../config", "./output.xyz")
 	#process_genome(0,2,[0,8,0],"/alcc/gpfs2/home/u/blaschma/rotating_structures")
-	process_genome(0,8,[0, 0, 0],"/alcc/gpfs2/home/u/blaschma/relax_test")
+	process_genome(0,1,[0,0,0,6,0,2,0],"/alcc/gpfs2/home/u/blaschma/genetic_runs/genetic_run_31/test/")
 
 
 	"""
